decoding
All the subplots having ‘frequency (Hz)’ as x-axis, ‘|Y(f)|’ as y-axis. I hided the x and y labels here in order to obtain a clearer plot. All the frequencies contained in each subplot is recorded in the table below. The order of subplots going from left to right, there are 64 subplots in total since the audio file is 64 seconds in total.